Fixed messages path
