+++ port/js/package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "build": "node_modules/.bin/rollup -c",
     "test": "jest --coverage",
-    "gen-json": "python3 ../../generate.py -b ../../tests/ -m ./messages/messgen/messgen_test -l json -o ../../tests/js",
-    "gen-cpp": "python3 ../../generate.py -b ../../tests/ -m ./messages/messgen/messgen_test -l cpp -o ../../tests/js"
+    "gen-json": "python3 ../../generate.py -b ../../tests/messages -m messgen/messgen_test -l json -o ../../tests/js/gen",
+    "gen-cpp": "python3 ../../generate.py -b ../../tests/messages -m messgen/messgen_test -l cpp -o ../../tests/js/gen"
   },
   "author": "",
   "license": "ISC",
