+++ tests/js/gen/messgen/msgs/messgen_test/messages.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <messgen/msgs/messgen_test/simple_message.h>
+#include <messgen/msgs/messgen_test/embedded_message_d1.h>
+#include <messgen/msgs/messgen_test/embedded_message_d2.h>
+#include <messgen/msgs/messgen_test/simple_dynamic_message.h>
+#include <messgen/msgs/messgen_test/embedded_dynamic_message_d1.h>
+#include <messgen/msgs/messgen_test/empty.h>
+#include "proto.h"
+#include "constants.h"
+
