+++ tests/js/main.cpp
@@ -3,8 +3,7 @@ using namespace std;
 #include "messgen/messgen.h"
 #include "messgen/stl.h"
 
-//#include "simple_message.h"
-
+#include <messgen/msgs/messgen_test/messages.h>
 
 #include <iostream>
 #include <vector>
