--- tests/js/msgs/messages/messgen/messgen_test/messages.h
@@ -1,11 +1,0 @@
-#pragma once
-
-#include <./msgs/messages/messgen/messgen_test/simple_message.h>
-#include <./msgs/messages/messgen/messgen_test/embedded_message_d1.h>
-#include <./msgs/messages/messgen/messgen_test/embedded_message_d2.h>
-#include <./msgs/messages/messgen/messgen_test/simple_dynamic_message.h>
-#include <./msgs/messages/messgen/messgen_test/embedded_dynamic_message_d1.h>
-#include <./msgs/messages/messgen/messgen_test/empty.h>
-#include "proto.h"
-#include "constants.h"
-
